Use base handler for index email job
